+++ src/Container/PublisherCommandFactory.php
@@ -2,10 +2,10 @@
 
 namespace ZfrEbWorker\Container;
 
-use Aws\Sdk as AwsSdk;
 use Interop\Container\ContainerInterface;
 use ZfrEbWorker\Cli\PublisherCommand;
-use ZfrEbWorker\Publisher\QueuePublisherInterface;
+use ZfrEbWorker\MessageQueue\InMemoryMessageQueueRepository;
+use ZfrEbWorker\MessageQueue\MessageQueueRepositoryInterface;
 
 /**
  * @author Michaël Gallego
@@ -18,13 +18,9 @@ class PublisherCommandFactory
      */
     public function __invoke(ContainerInterface $container): PublisherCommand
     {
-        /** @var QueuePublisherInterface $queuePublisher */
-        $queuePublisher = $container->get(QueuePublisherInterface::class);
+        /** @var InMemoryMessageQueueRepository $queueRepository */
+        $queueRepository = $container->get(MessageQueueRepositoryInterface::class);
 
-        /** @var AwsSdk $awsSdk */
-        $awsSdk    = $container->get(AwsSdk::class);
-        $sqsClient = $awsSdk->createSqs();
-
-        return new PublisherCommand($queuePublisher, $sqsClient);
+        return new PublisherCommand($queueRepository);
     }
 }