+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 4.0.0
+
+* `QueuePublisherInterface` has been removed and replaced by a single `MessageQueueInterface`, tied to a single queue.
+* A new `MessageQueueRepositoryInterface` has been introduced to allow fetching message queues by name.
+* Added a new `Message` and `DelayedMessage` abstraction for easier creation of messages.
+
 # 3.3.0
 
 * Add a new `SilentFailingListener` class that you can use if you don't want to handle a specific message.